Cast it.

+++ trunk/src/linad99/def_eval.cpp
@@ -188,8 +188,8 @@
     {
       // \todo Need test
       // back up the file one buffer size and read forward
-      off_t offset = sizeof(grad_stack_entry)
-        * gradient_structure::GRAD_STACK1->length;
+      off_t offset = (off_t)(sizeof(grad_stack_entry)
+        * gradient_structure::GRAD_STACK1->length);
       off_t lpos=lseek(gradient_structure::GRAD_STACK1->_GRADFILE_PTR,
         -offset, SEEK_CUR);
 
@@ -223,8 +223,8 @@
     {
       // \todo Need test
       // back up the file one buffer size and read forward
-      off_t offset = sizeof(grad_stack_entry)
-        * gradient_structure::GRAD_STACK1->length;
+      off_t offset = (off_t)(sizeof(grad_stack_entry)
+        * gradient_structure::GRAD_STACK1->length);
       off_t lpos=lseek(gradient_structure::GRAD_STACK1->_GRADFILE_PTR,
         -offset, SEEK_CUR);
